Extract neptune_query from neptune_fetcher.v1
