+++ tests/e2e/v1/runs/test_runs_list_attributes.py
@@ -2,8 +2,8 @@ import os
 
 import pytest
 
-import neptune_fetcher.v1.runs as runs
-from neptune_fetcher.v1.filters import (
+import neptune_query.runs as runs
+from neptune_query.filters import (
     Attribute,
     AttributeFilter,
     Filter,
